# improve Array.permutations() performance #1995