@jest I believe that the Narayana algorithm is simple and fast enough