@jest 2017-05-18
@ruslansennov Probably it is, although "fast enough" is always usage-dependant.

I thought more general about S-T-J algorithm also in terms of memory reusing (e.g. as in Vector, where swap of two adjacent elements leaves most of trie branches untouched), but after rethinking, nobody will run `permutations()` on sequences long enough for storage sharing to make sense.

I'll check how this PR performs later in the day.