@jest 2017-05-18
master ce6f975
```
Benchmark                                       Mode  Cnt    Score    Error  Units
VavrPermutationTest.bench_dummy_no_permutation  avgt   20    2.417 ±  0.037  ms/op
VavrPermutationTest.bench_guava_permutation     avgt   20    2.699 ±  0.022  ms/op
VavrPermutationTest.bench_vavr_permutation      avgt   20  519.320 ± 10.301  ms/op
```

PR 403b7fc
```
Benchmark                                       Mode  Cnt  Score   Error  Units
VavrPermutationTest.bench_dummy_no_permutation  avgt   20  2.454 ± 0.022  ms/op
VavrPermutationTest.bench_guava_permutation     avgt   20  2.517 ± 0.016  ms/op
VavrPermutationTest.bench_vavr_permutation      avgt   20  1.791 ± 0.018  ms/op
```

Hm, this is fast, suspiciously fast...

So I rerun the test with 9 elements in the array:

```
Benchmark                                       Mode  Cnt   Score   Error  Units
VavrPermutationTest.bench_dummy_no_permutation  avgt   20  28.672 ± 0.361  ms/op
VavrPermutationTest.bench_guava_permutation     avgt   20  25.231 ± 0.383  ms/op
VavrPermutationTest.bench_vavr_permutation      avgt   20  17.353 ± 0.164  ms/op
```

And with 10 elements in the array (large variance is probably to GC, not tuned JMH in this aspect in any way):

```
Benchmark                                       Mode  Cnt     Score     Error  Units
VavrPermutationTest.bench_dummy_no_permutation  avgt   20  1444.303 ±  89.729  ms/op
VavrPermutationTest.bench_guava_permutation     avgt   20   848.617 ±  99.800  ms/op
VavrPermutationTest.bench_vavr_permutation      avgt   20   327.846 ± 245.760  ms/op
```

So if the result of permutation is correct, it is indeed much faster in this PR.