@jest 2017-05-18
Rerun the last test, within more stabilized environment:

```
VavrPermutationTest.bench_dummy_no_permutation  avgt   20  1519.123 ± 81.149  ms/op
VavrPermutationTest.bench_guava_permutation     avgt   20   776.917 ± 43.621  ms/op
VavrPermutationTest.bench_vavr_permutation      avgt   20   220.733 ±  2.601  ms/op
```

The further, the merrier! :)