@ruslansennov 2017-05-18
> So if the result of permutation is correct, it is indeed much faster in this PR.

It is correct, I checked this, trust me :)
Thank you for measurements