@danieldietrich 2017-05-18
@ruslansennov It would be good to move the `distinct()` to the 3rd if branch, where it is needed. There are some instances created otherwise, even if the Array is empty:

```java
    @Override
    public Array<T> distinct() {
        return distinctBy(Function.identity());
    }

    @Override
    public <U> Array<T> distinctBy(Function<? super T, ? extends U> keyExtractor) {
        Objects.requireNonNull(keyExtractor, "keyExtractor is null");
        final java.util.Set<U> seen = new java.util.HashSet<>();
        return filter(t -> seen.add(keyExtractor.apply(t)));
    }
```