@ruslansennov 2017-05-18
> There are some instances created otherwise, even if the Array is empty

OK, your're right! I'll do it this evening