@ruslansennov 2017-05-18
I added test for permutation of repeated elements and it seems that `lexicographic order` (both: `S-T-J` and `Narayana`) is bad choice in our case. These algorithms do not take into account repeating of elements and produce result different from other Vavr collections (as well as Scala collections) because they operate with indexes but not with elements. For example:
```java
a1 = [1, 1]
Narayana permutations: [[1,1], [1,1]]
Narayana permutations of distinct elements: [[1]]
Other Vavr collections and Scala: [[1,1]]

a1 = [1, 2, 2]
Narayana permutations: [[1, 2, 2], [1, 2, 2], [2, 1, 2], [2, 2, 1], [2, 1, 2], [2, 2, 1]]
Narayana permutations of distinct elements: [[1, 2], [2, 1]]
Other Vavr collections and Scala: [[1, 2, 2], [2, 1, 2], [2, 2, 1]]
```
So sad, but I don't see any solution of this problem and we should revert all changes to very first algorithm (v.2.0.6). It will be not so fast, but right in all cases