@ruslansennov 2017-05-18
Or, we can return `Narayana.distinct()` sequence as result of permutations, but this will be very slow for big `N`