@danieldietrich 2017-05-19
It does also not help to first calculate the Narayana permutations and then call distinct():

```
[[1, 2, 2], [1, 2, 2], [2, 1, 2], [2, 2, 1], [2, 1, 2], [2, 2, 1]].distinct() == [[1, 2, 2], [2, 1, 2], [2, 2, 1]]
```

For big `N` it will perform not well. The distinct() has to be called first. However, as you said, the result is different.

Let's take the first algorithm for now.

Any other suggestions are welcome.