@jest 2017-05-19
Hm, I'm not sure, probably Narayana's can handle the repeated elements (http://www.mafagafogigante.org/lexicographic-permutations/), when the "grater-or-equal" is  used in `while` loops. Of course we are working on indices instead of the elements, but with additional array to `pi` used for comparisons it should work.

E.g. for `[a, b, b, c]` we would have `pi = [0, 1, 2, 3]` used for permutation generation and `ci = [0, 1, 1, 2]` used for comparisons in `while`-s. And swapping of course elements of both `pi` and `ci` simultaneously. I can try it on the weekend.

Anyway, this is IMO a breaking change in the API?