@danieldietrich 2017-05-19
@jest 

> Anyway, this is IMO a breaking change in the API?

If the newly added unit test #1998 works, it will be fine.

> I can try it on the weekend.

Any helping hands are welcome :)