@jest 2017-05-19
One question: is the order of permutations in result an implementation detail, a "nice-to-have" feature (e.g. lexicographic order) or part of API?

I'm not sure, I'll be easily able to handle repeated elements which are not grouped, like:

```
arr = [b, a, c, a, b]
pi=[0, 1, 2, 3, 4]
ci=[0, 1, 2, 1, 0]
```

Because `ci` is not first in the lexicographical order (of positions, as `pi` is) and I'll never get to permutation `[b, b, a, a, c]` (`ci=[0, 0, 1, 1, 2]`); hence, I can only generate lexicographical order in the sense of `ci` order, by first  grouping repeating elements, like `[b, a, c, a, b] => [b, b, a, a, c]` and generate permutation of that.