@danieldietrich 2017-05-19
@jest This reminds me of java.util.Map. A few years ago some users relied on the keySet() element order, even if the API states that it is unpredictable. Then Java changed its internal implementation and applications broke.

The question is, do we expect Seq.permutations() to behave the same along all Seq implementations?
Following the Liskov Substitution Principle I would say yes, unless the Seq.permutations() API states the order is unpredictable.

It is always a good choice to make it clear using types. E.g. returning a `Set<? extends Seq<T>>` of permutations instead of a collection of the same type, e.g. `Array<Array<T>>`.

However, Scala does return an `Iterator`. Because computing permutations is costly (time & mem), this is a good idea. What if we do the same and provide a permutation iterator with a consistent order along all Seq implementations? We could do this change on a 1.0.0 branch and provide the old implementation on master that will get the version 0.9.1.